clean up layout context
