+++ src/engraving/layout/layoutcontext.h
@@ -22,57 +22,48 @@
 #ifndef MU_ENGRAVING_LAYOUTCONTEXT_H
 #define MU_ENGRAVING_LAYOUTCONTEXT_H
 
-#include "libmscore/score.h"
-#include "libmscore/page.h"
 #include "libmscore/fraction.h"
-#include "libmscore/system.h"
-#include "libmscore/spanner.h"
+
+namespace Ms {
+class Score;
+class Page;
+class System;
+class Spanner;
+class MeasureBase;
+}
 
 namespace mu::engraving {
-//---------------------------------------------------------
-//   LayoutContext
-//    temp values used during layout
-//---------------------------------------------------------
-
-class LayoutContext
+struct LayoutContext
 {
-public:
+    Ms::Score* score = nullptr;
+    bool startWithLongNames = true;
+    bool firstSystem = true;
+    bool firstSystemIndent = true;
+    Ms::Page* page = nullptr;
+    int curPage = 0; // index in Score->page()s
+    Ms::Fraction tick{ 0, 1 };
+
+    QList<Ms::System*> systemList; // reusable systems
+    std::set<Ms::Spanner*> processedSpanners;
+
+    Ms::System* prevSystem = nullptr; // used during page layout
+    Ms::System* curSystem = nullptr;
+
+    Ms::MeasureBase* systemOldMeasure = nullptr;
+    Ms::MeasureBase* pageOldMeasure = nullptr;
+    bool rangeDone = false;
+
+    Ms::MeasureBase* prevMeasure = nullptr;
+    Ms::MeasureBase* curMeasure = nullptr;
+    Ms::MeasureBase* nextMeasure = nullptr;
+    int measureNo = 0;
+    Ms::Fraction startTick;
+    Ms::Fraction endTick;
+
     LayoutContext(Ms::Score* s);
     LayoutContext(const LayoutContext&) = delete;
     LayoutContext& operator=(const LayoutContext&) = delete;
     ~LayoutContext();
-
-    void layoutLinear();
-
-    void layout();
-    int adjustMeasureNo(Ms::MeasureBase*);
-    void getNextPage();
-    void collectPage();
-
-    Ms::Score* score             { 0 };
-    bool startWithLongNames  { true };
-    bool firstSystem         { true };
-    bool firstSystemIndent   { true };
-    Ms::Page* page               { 0 };
-    int curPage              { 0 };        // index in Score->page()s
-    Ms::Fraction tick            { 0, 1 };
-
-    QList<Ms::System*> systemList;            // reusable systems
-    std::set<Ms::Spanner*> processedSpanners;
-
-    Ms::System* prevSystem       { 0 };       // used during page layout
-    Ms::System* curSystem        { 0 };
-
-    Ms::MeasureBase* systemOldMeasure { 0 };
-    Ms::MeasureBase* pageOldMeasure   { 0 };
-    bool rangeDone           { false };
-
-    Ms::MeasureBase* prevMeasure { 0 };
-    Ms::MeasureBase* curMeasure  { 0 };
-    Ms::MeasureBase* nextMeasure { 0 };
-    int measureNo            { 0 };
-    Ms::Fraction startTick;
-    Ms::Fraction endTick;
 };
 }
